Move classes closer to where they belong.
