--- vavr/src/test/java/io/vavr/collection/IntegerValue.java
@@ -1,76 +1,0 @@
-/*  __    __  __  __    __  ___
- * \  \  /  /    \  \  /  /  __/
- *  \  \/  /  /\  \  \/  /  /
- *   \____/__/  \__\____/__/
- *
- * Copyright 2014-2017 Vavr, http://vavr.io
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.vavr.collection;
-
-import java.util.function.BiFunction;
-
-public class IntegerValue {
-    private static final IntegerValueMonoid integerValueMonoid = new IntegerValueMonoid();
-
-    private final int integerValue;
-
-    public IntegerValue(int integerValue) {
-        this.integerValue = integerValue;
-    }
-
-    public static IntegerValue with(int integerValue) {
-        return new IntegerValue(integerValue);
-    }
-
-    public IntegerValue add(IntegerValue that) {
-        return new IntegerValue(this.integerValue + that.integerValue);
-    }
-
-    public static IntegerValueMonoid monoid() {
-        return integerValueMonoid;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other instanceof IntegerValue) {
-            IntegerValue that = (IntegerValue) other;
-            return this.integerValue == that.integerValue;
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public int hashCode() {
-        return integerValue;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("IntegerValue[integerValue=%d]", integerValue);
-    }
-
-    public static class IntegerValueMonoid implements Monoid<IntegerValue> {
-        @Override
-        public IntegerValue identityElement() {
-            return with(0);
-        }
-
-        @Override
-        public BiFunction<IntegerValue, IntegerValue, IntegerValue> addFunction() {
-            return IntegerValue::add;
-        }
-    }
-}
